backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.0.3
SQLAlchemy==2.0.25
asyncpg==0.29.0
aiosqlite==0.19.0
alembic==1.12.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
email-validator==2.1.0.post1
